Dwite now detects (through an application level watchdog) if the device becomes unreachable on the network (e.g. because it powered down). If so, the Dwite main() loop simply starts over, allowing the device to connect again.
The watchdog has two timeouts: One for when the device is considered
timed out and dead, and a shorter "sleep" timeout during which the
device is not supposed to be polled. The watchdog is reset 50 times
per second by the device handler and it doesn't make sense to send
a new polling message over the wire each time.

The polling message is implemented using the "stat" command without
any parameters. It will cause the device to answer with a regular
STAT message where .event="stat". There is no documentation in SBS
sources about the "stat" command, but it is pretty obvious what it
does: Ask the device to send a STAT message.